最新版本freqccv4-adaptive，在数据流速率较低时排队延迟显著差于hybrid

+++ NS3.27/src/dqc/model/dqc_receiver.h
@@ -78,6 +78,6 @@ private:
     uint32_t m_timeLastPacket=0;
     uint64_t m_sumOwd=0;
     uint32_t m_lastCalGoodTime=0;
-    uint32_t m_lastGoodRecv=0;
+    uint64_t m_lastGoodRecv=0;
 };    
 }
